move expiration functions to utils
