+++ reconcile/openshift_groups.py
@@ -6,7 +6,7 @@ from sretoolbox.utils import threaded
 
 from reconcile.utils import gql
 from reconcile import queries
-from reconcile import openshift_rolebindings
+from reconcile.utils import expiration
 from reconcile.utils.oc import OC_Map
 from reconcile.utils.defer import defer
 from reconcile.utils.sharding import is_in_shard
@@ -114,13 +114,12 @@ def fetch_desired_state(oc_map):
     desired_state = []
 
     for r in roles:
-        if not openshift_rolebindings \
-                .has_valid_expiration_date(r['expirationDate']):
+        if not expiration.has_valid_expiration_date(r['expirationDate']):
             raise ValueError(
                 f'expirationDate field is not formatted as YYYY-MM-DD, '
                 f'currently set as {r["expirationDate"]}'
             )
-        if not openshift_rolebindings.role_still_valid(r['expirationDate']):
+        if not expiration.role_still_valid(r['expirationDate']):
             logging.warning(
                 f'The maximum expiration date of {r["name"]} '
                 f'has passed '
